Frontend i18n, silent JWT refresh, section-heading styling, ROUND fix
Frontend (app/static/index.html):
- Localize the post-upload starter chips to French when the language
  dropdown is FR — category titles, chip labels, and the prompts sent
  to the agent all switch language. _onLanguageChange() re-renders the
  banner without wiping any in-progress conversation.
- Silent JWT refresh inspired by the keycloak-js axios-interceptor
  pattern, ported to vanilla JS: _decodeJwtPayload(), _tokenExpSec(),
  _isTokenNearExpiry(), _refreshAuthToken() (single-flight),
  _ensureFreshToken(), _apiFetch() wrapper, plus a 60-second background
  heartbeat that refreshes when the token is within 5 minutes of expiry.
  Wired into initAuth() (eager refresh on restore, then start timer)
  and logOut() (stop timer).
- Fix section-heading styling for assistant responses: the previous
  detector only matched **Bold** form with a short name list, so the
  file-upload agent's `## Réponse` / `## Analyse` /
  `## Aperçus et prochaine étape` rendered as plain bold instead of
  the orange section-header treatment. New sectionHdrLabel() matches
  both `**X**` and `## X` against an expanded name list covering every
  heading the backend emits in EN and FR.

Backend (app/api/auth_routes.py):
- New POST /api/v1/auth/refresh endpoint. Takes a valid Bearer JWT,
  returns a fresh 7-day JWT with the SAME session_token (sid). We do
  not rotate session_token on refresh — rotation stays reserved for
  /login and /keycloak/callback (where it's intentional one-active-
  device behavior). Refreshing in one tab won't 401 another tab.

Bug fix (app/agents/monthly_report.py):
- Wrap the two ROUND() CASE expressions in ::numeric casts. Postgres
  ROUND(value, digits) only exists for numeric; arithmetic on
  double-precision columns gives double, which has no two-arg ROUND.
  Same pattern is already followed elsewhere in the file; these two
  were oversights.

+++ app/agents/monthly_report.py
@@ -80,13 +80,13 @@ def _fetch_data() -> dict[str, Any]:
                 ROUND(d.budget_value::numeric,  0)       AS budget,
                 ROUND(d.last_year_real_value::numeric, 0) AS prior_year,
                 CASE WHEN d.budget_value IS NOT NULL AND d.budget_value <> 0
-                     THEN ROUND((d.real_value - d.budget_value) * 100.0
-                                / ABS(d.budget_value), 1)
+                     THEN ROUND(((d.real_value - d.budget_value) * 100.0
+                                 / ABS(d.budget_value))::numeric, 1)
                      ELSE NULL END                       AS vs_budget_pct,
                 CASE WHEN d.last_year_real_value IS NOT NULL
                           AND d.last_year_real_value <> 0
-                     THEN ROUND((d.real_value - d.last_year_real_value) * 100.0
-                                / ABS(d.last_year_real_value), 1)
+                     THEN ROUND(((d.real_value - d.last_year_real_value) * 100.0
+                                 / ABS(d.last_year_real_value))::numeric, 1)
                      ELSE NULL END                       AS yoy_pct
         FROM    dedup d
         JOIN    financial_metric     fm ON fm.id = d.financial_metric_id
